STSMACOM-227: Add tests.
