+++ lib/EditableList/processBadResponse.js
@@ -54,6 +54,15 @@ const getErrorMessages = (errors, defaultErrorCode) => {
   return errorMessages;
 };
 
+const getDefaultErrorMessage = (defaultErrorCode) => {
+  const message = getErrorMessage(defaultErrorCode);
+
+  return {
+    fieldErrors: {},
+    commonErrors: [message],
+  };
+};
+
 export default async function processBadResponse(response, defaultErrorCode) {
   if (response.status === 422) {
     try {
@@ -61,9 +70,9 @@ export default async function processBadResponse(response, defaultErrorCode) {
 
       return getErrorMessages(errors, defaultErrorCode);
     } catch (e) {
-      return [getErrorMessage(defaultErrorCode)];
+      return getDefaultErrorMessage(defaultErrorCode);
     }
   }
 
-  return [getErrorMessage(defaultErrorCode)];
+  return getDefaultErrorMessage(defaultErrorCode);
 }
