+++ tests/network/scenarios/controlled-vocab-custom-error.js
@@ -1,0 +1,12 @@
+export default function (server) {
+  server.post('location-units/institutions', {
+    'errors': [{
+      'message': 'test',
+      'code': '-1',
+      'parameters': [{
+        'key': 'test',
+        'value': 'test'
+      }]
+    }]
+  }, 422);
+}
