+++ tests/network/scenarios/controlled-vocab-default-error.js
@@ -1,0 +1,12 @@
+export default function (server) {
+  server.post('location-units/institutions', {
+    'errors': [{
+      'message': 'Cannot create entity; name is not unique',
+      'code': 'name.duplicate',
+      'parameters': [{
+        'key': 'fieldLabel',
+        'value': 'name'
+      }]
+    }]
+  }, 400);
+}
